Changes the project structure in order to improve code management
