+++ BudgetManager/utils/app_authentication/ConfirmationSender.cs
