+++ BudgetManager/utils/app_authentication/PasswordSecurityManager.cs
