+++ BudgetManager/utils/data_insertion/DataInsertionCheckStrategy.cs
