+++ BudgetManager/utils/data_insertion/DataInsertionCheckerContext.cs
