+++ BudgetManager/utils/data_insertion/DataInsertionStrategy.cs
