+++ BudgetManager/utils/data_insertion/ExpenseInsertionStrategy.cs
@@ -1,5 +1,6 @@
 ﻿using BudgetManager.mvc.models.dto;
 using BudgetManager.utils;
+using BudgetManager.utils.enums;
 using MySql.Data.MySqlClient;
 using System;
 using System.Collections.Generic;
