+++ BudgetManager/utils/data_insertion/IncomeInsertionStrategy.cs
