+++ BudgetManager/utils/data_insertion/ReceivableInsertionStrategy.cs
