+++ BudgetManager/utils/data_insertion/TransferCheckStrategy.cs
