+++ BudgetManager/utils/data_transfer/QueryData.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using BudgetManager.utils.enums;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
