+++ BudgetManager/utils/enums/IncomeSource.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace BudgetManager.utils.enums {
+    public enum IncomeSource {
+        GENERAL_INCOMES,
+        SAVING_ACCOUNT,
+        UNDEFINED
+    }
+}
