+++ BudgetManager/utils/miscellaneous/ViewCalculator.cs
