+++ BudgetManager/utils/ui_controls/UserControlsManager.cs
